+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAB1TGN47491SVA9STS1GEM4"
+        "01KABAYG2CXRD08HBXHEK5625B"
     )
     r"""The unique identifier of the data source"""