+++ packages/orq-rc/src/orq_ai_sdk/models/duplicatetoolop.py
@@ -204,7 +204,7 @@ class DuplicateToolResponseBody5(BaseModel):
     code_tool: DuplicateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGHVWCMVPEPMBHXER7ZR"
+        "tool_01KABAYFYN29VA2VX30PZ5G11Z"
     )
 
     display_name: Optional[str] = None
@@ -278,7 +278,7 @@ class DuplicateToolResponseBodyTools(BaseModel):
         DuplicateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KAB1TGHTXR8PJ2MJVC5XC435"
+    id: Optional[str] = "01KABAYFYMV0NCGDD4EM1NWJ7G"
 
     description: Optional[str] = None
 
@@ -373,7 +373,7 @@ class DuplicateToolResponseBody4(BaseModel):
     mcp: DuplicateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGHS1F81DNVYVYMJS1BS"
+        "tool_01KABAYFYKXVMRYBBKXXGC6DN0"
     )
 
     display_name: Optional[str] = None
@@ -581,7 +581,7 @@ class DuplicateToolResponseBody3(BaseModel):
     http: DuplicateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGHQHV6QETE29NB6B937"
+        "tool_01KABAYFYHRS6JHAR63T5CZ0GH"
     )
 
     display_name: Optional[str] = None
@@ -731,7 +731,7 @@ class DuplicateToolResponseBody2(BaseModel):
     json_schema: DuplicateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGHPVCQGZDNKJSM8SCMQ"
+        "tool_01KABAYFYG896ZS9JBZ6E2V5JT"
     )
 
     display_name: Optional[str] = None
@@ -885,7 +885,7 @@ class DuplicateToolResponseBody1(BaseModel):
     function: DuplicateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGHNM52QRHY0FHKXMTRJ"
+        "tool_01KABAYFYEPEEP6QH2HDC5FG63"
     )
 
     display_name: Optional[str] = None