+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAB1TGN0G7NZJ4X01N6N32G6"
+        "01KABAYG28YN4GRZDGFFM78APA"
     )
     r"""The unique identifier of the data source"""