+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -188,7 +188,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KAB1TGHBRY9FFWZN5BBXGRDP"
+    id: Optional[str] = "01KABAYFY3PFP54H7MC3P1ZX48"
 
     description: Optional[str] = None
 
@@ -907,7 +907,7 @@ class UpdateToolResponseBody5(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGH2RVKAYQ5WM8W41X7J"
+        "tool_01KABAYFXT3KEN25GJ4TX40R2R"
     )
 
     display_name: Optional[str] = None
@@ -981,7 +981,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KAB1TGH1NYW0M0C1QS1921S5"
+    id: Optional[str] = "01KABAYFXS8119Q9NPMWZ75PS1"
 
     description: Optional[str] = None
 
@@ -1076,7 +1076,7 @@ class UpdateToolResponseBody4(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGH0B9D7XAS0T2MXF05Q"
+        "tool_01KABAYFXRHJ9B6737RCJP0FJR"
     )
 
     display_name: Optional[str] = None
@@ -1284,7 +1284,7 @@ class UpdateToolResponseBody3(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGGYNPXFCDXHMJQEG3ET"
+        "tool_01KABAYFXP9997QRF6DDW8WXZ2"
     )
 
     display_name: Optional[str] = None
@@ -1434,7 +1434,7 @@ class UpdateToolResponseBody2(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGGWGX7RV8YYFGJ2PT8T"
+        "tool_01KABAYFXM62KC83SVCBZ2B1S6"
     )
 
     display_name: Optional[str] = None
@@ -1588,7 +1588,7 @@ class UpdateToolResponseBody1(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KAB1TGGVZP2BGHHS0EVP4NSR"
+        "tool_01KABAYFXKAMPRTKCWMKNAA8EK"
     )
 
     display_name: Optional[str] = None