+++ packages/orq-rc/src/orq_ai_sdk/tools.py
@@ -3,6 +3,10 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
+from orq_ai_sdk.models import (
+    createtoolop as models_createtoolop,
+    updatetoolop as models_updatetoolop,
+)
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -210,7 +214,10 @@ class Tools(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[models.CreateToolRequestBody, models.CreateToolRequestBodyTypedDict]
+            Union[
+                models_createtoolop.CreateToolRequestBody,
+                models_createtoolop.CreateToolRequestBodyTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -307,7 +314,10 @@ class Tools(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[models.CreateToolRequestBody, models.CreateToolRequestBodyTypedDict]
+            Union[
+                models_createtoolop.CreateToolRequestBody,
+                models_createtoolop.CreateToolRequestBodyTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -405,7 +415,10 @@ class Tools(BaseSDK):
         *,
         tool_id: str,
         request_body: Optional[
-            Union[models.UpdateToolRequestBody, models.UpdateToolRequestBodyTypedDict]
+            Union[
+                models_updatetoolop.UpdateToolRequestBody,
+                models_updatetoolop.UpdateToolRequestBodyTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -513,7 +526,10 @@ class Tools(BaseSDK):
         *,
         tool_id: str,
         request_body: Optional[
-            Union[models.UpdateToolRequestBody, models.UpdateToolRequestBodyTypedDict]
+            Union[
+                models_updatetoolop.UpdateToolRequestBody,
+                models_updatetoolop.UpdateToolRequestBodyTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,